refactor: 패키지 구조 변경, BaseEntity 상속 및 Entity 수정
