+++ src/main/java/org/boot/growup/order/persist/repository/DeliveryRepository.java
@@ -1,6 +1,6 @@
-package org.boot.growup.source.order.persist.repository;
+package org.boot.growup.order.persist.repository;
 
-import org.boot.growup.source.order.persist.entity.Delivery;
+import org.boot.growup.order.persist.entity.Delivery;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface DeliveryRepository extends JpaRepository<Delivery, Long> {
