+++ src/main/java/org/boot/growup/order/persist/repository/OrderCancelRepository.java
@@ -1,6 +1,6 @@
-package org.boot.growup.source.order.persist.repository;
+package org.boot.growup.order.persist.repository;
 
-import org.boot.growup.source.order.persist.entity.OrderCancel;
+import org.boot.growup.order.persist.entity.OrderCancel;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface OrderCancelRepository extends JpaRepository<OrderCancel, Long> {
